使用realm，处理关联表更方便，待调试。

+++ Vodka/Manager/DLFeedFetcher.h
@@ -20,7 +20,7 @@
 
 
 //从数据库中分页取出feeds
--(void)fetchItems:(NSInteger)offset limit:(NSInteger)limit completion:(void (^)(NSArray <DLFeedItem *>*feedItems))completion;
+-(void)fetchItems:(int)offset limit:(int)limit completion:(void (^)(NSArray <DLFeedItem *>*feedItems))completion;
 
 //解析feeds，并存入数据库
 -(void)loadFeeds;
